Tabular data view functionality gui
